+++ simulation/python/lockin/__init__.py
@@ -1,0 +1,11 @@
+from .dds import DDS, MultitoneDDS
+from .filters import CICDecimator, FIRFilter, design_lowpass_fir, cic_frequency_response
+from .demod import DemodChannel, MultiToneDemodulator
+from .chain import LockInChain, LockInChainParams, TIAParams, ADCParams
+
+__all__ = [
+    "DDS", "MultitoneDDS",
+    "CICDecimator", "FIRFilter", "design_lowpass_fir", "cic_frequency_response",
+    "DemodChannel", "MultiToneDemodulator",
+    "LockInChain", "LockInChainParams", "TIAParams", "ADCParams",
+]